Fix request changes #1 and fix line chart
