+++ src/api/interopAxiosClient.tsx
@@ -1,6 +1,6 @@
 import { AxiosInstance, AxiosResponse } from 'axios'
 import { createAxiosInstance } from './axiosInstanceCreator'
-import { getServicesType, getServicesTelemetry } from './apiRequestTypes'
+import { GetServicesType, GetServicesTelemetry } from './apiRequestTypes'
 
 class Http {
   private instance: AxiosInstance | null = null
@@ -9,7 +9,7 @@ class Http {
     return this.instance != null ? this.instance : createAxiosInstance()
   }
 
-  getServices<T = any, R = AxiosResponse<T>>(payload: getServicesType): Promise<R> {
+  getServices<T = any, R = AxiosResponse<T>>(payload: GetServicesType): Promise<R> {
     return this.http.get<T, R>('/eservices', {
       params: {
         offset: payload.offset,
@@ -42,7 +42,7 @@ class Http {
   }
 
   getServiceStatisticsData<T = any, R = AxiosResponse<T>>(
-    payload: getServicesTelemetry
+    payload: GetServicesTelemetry
   ): Promise<R> {
     return this.http.get<T, R>(`/eservices/${payload.eserviceRecordId}`, {
       params: {
